style: format all files

+++ contracts/vesting/Genesis.sol
@@ -16,17 +16,17 @@ interface IClaim {
         uint256 wClaimed; // rebase-tracking number
         uint256 max; // maximum nominal OHM amount can claim
     }
+
     function terms(address _address) external view returns (Term memory);
 }
 
 /**
  *  This contract allows Olympus genesis contributors to claim OHM. It has been
  *  revised to consider 9/10 tokens as staked at the time of claim; previously,
- *  no claims were treated as staked. This change keeps network ownership in check. 
+ *  no claims were treated as staked. This change keeps network ownership in check.
  *  100% can be treated as staked, if the DAO sees fit to do so.
  */
 contract GenesisClaim is Ownable {
-
     /* ========== DEPENDENCIES ========== */
 
     using SafeMath for uint256;
@@ -44,15 +44,15 @@ contract GenesisClaim is Ownable {
     /* ========== STATE VARIABLES ========== */
 
     // claim token
-    IERC20 internal immutable ohm = IERC20(0x64aa3364F17a4D01c6f1751Fd97C2BD3D7e7f1D5); 
+    IERC20 internal immutable ohm = IERC20(0x64aa3364F17a4D01c6f1751Fd97C2BD3D7e7f1D5);
     // payment token
-    IERC20 internal immutable dai = IERC20(0x6B175474E89094C44Da98b954EedeAC495271d0F); 
+    IERC20 internal immutable dai = IERC20(0x6B175474E89094C44Da98b954EedeAC495271d0F);
     // mints claim token
-    ITreasury internal immutable treasury = ITreasury(0x9A315BdF513367C0377FB36545857d12e85813Ef); 
+    ITreasury internal immutable treasury = ITreasury(0x9A315BdF513367C0377FB36545857d12e85813Ef);
     // stake OHM for sOHM
-    IStaking internal immutable staking = IStaking(0xB63cac384247597756545b500253ff8E607a8020); 
+    IStaking internal immutable staking = IStaking(0xB63cac384247597756545b500253ff8E607a8020);
     // holds non-circulating supply
-    address internal immutable dao = 0x245cc372C84B3645Bf0Ffe6538620B04a217988B; 
+    address internal immutable dao = 0x245cc372C84B3645Bf0Ffe6538620B04a217988B;
     // tracks rebase-agnostic balance
     IgOHM internal immutable gOHM = IgOHM(0x0ab87046fBb341D058F17CBC4c1133F25a20a52f);
     // previous deployment of contract (to migrate terms)
@@ -61,18 +61,20 @@ contract GenesisClaim is Ownable {
     // track 1/10 as static. governance can disable if desired.
     bool public useStatic;
     // tracks address info
-    mapping( address => Term ) public terms;
+    mapping(address => Term) public terms;
     // facilitates address change
-    mapping( address => address ) public walletChange;
+    mapping(address => address) public walletChange;
     // as percent of supply (4 decimals: 10000 = 1%)
     uint256 public totalAllocated;
     // maximum portion of supply can allocate. == 7.8%
-    uint256 public maximumAllocated = 78000; 
+    uint256 public maximumAllocated = 78000;
 
-    constructor() {useStatic = true;}
+    constructor() {
+        useStatic = true;
+    }
 
     /* ========== MUTABLE FUNCTIONS ========== */
-    
+
     /**
      * @notice allows wallet to claim OHM
      * @param _to address
@@ -89,7 +91,12 @@ contract GenesisClaim is Ownable {
      * @param _rebasing bool
      * @param _claimFromStaking bool
      */
-    function stake(address _to, uint256 _amount, bool _rebasing, bool _claimFromStaking) external {
+    function stake(
+        address _to,
+        uint256 _amount,
+        bool _rebasing,
+        bool _claimFromStaking
+    ) external {
         staking.stake(_to, _claim(_amount), _rebasing, _claimFromStaking);
     }
 
@@ -107,7 +114,7 @@ contract GenesisClaim is Ownable {
         require(redeemableFor(msg.sender).div(1e9) >= toSend_, "Claim more than vested");
         require(info.max.sub(claimed(msg.sender)) >= toSend_, "Claim more than max");
 
-        if(useStatic) {
+        if (useStatic) {
             terms[msg.sender].gClaimed = info.gClaimed.add(gOHM.balanceTo(toSend_.mul(9).div(10)));
             terms[msg.sender].claimed = info.claimed.add(toSend_.div(10));
         } else terms[msg.sender].gClaimed = info.gClaimed.add(gOHM.balanceTo(toSend_));
@@ -121,7 +128,7 @@ contract GenesisClaim is Ownable {
         require(terms[msg.sender].percent != 0, "No wallet to change");
         walletChange[msg.sender] = _newAddress;
     }
-    
+
     /**
      * @notice allows new address to pull terms
      * @param _oldAddress address
@@ -129,7 +136,7 @@ contract GenesisClaim is Ownable {
     function pullWalletChange(address _oldAddress) external {
         require(walletChange[_oldAddress] == msg.sender, "Old wallet did not push");
         require(terms[msg.sender].percent != 0, "Wallet already exists");
-        
+
         walletChange[_oldAddress] = address(0);
         terms[msg.sender] = terms[_oldAddress];
         delete terms[_oldAddress];
@@ -150,11 +157,11 @@ contract GenesisClaim is Ownable {
      * @param _address address
      * @return uint256
      */
-    function redeemableFor( address _address ) public view returns (uint) {
-        Term memory info = terms[ _address ];
-        uint max = circulatingSupply().mul( info.percent ).div( 1e6 );
+    function redeemableFor(address _address) public view returns (uint256) {
+        Term memory info = terms[_address];
+        uint256 max = circulatingSupply().mul(info.percent).div(1e6);
         if (max > info.max) max = info.max;
-        return max.sub( claimed( _address ) ).mul( 1e9 );
+        return max.sub(claimed(_address)).mul(1e9);
     }
 
     /**
@@ -184,13 +191,7 @@ contract GenesisClaim is Ownable {
     function migrate(address[] memory _addresses) external onlyOwner {
         for (uint256 i = 0; i < _addresses.length; i++) {
             IClaim.Term memory term = previous.terms(_addresses[i]);
-            setTerms(
-                _addresses[i], 
-                term.percent,
-                term.claimed,
-                term.wClaimed,
-                term.max
-            );
+            setTerms(_addresses[i], term.percent, term.claimed, term.wClaimed, term.max);
         }
     }
 
@@ -204,19 +205,14 @@ contract GenesisClaim is Ownable {
      *  @param _max uint256
      */
     function setTerms(
-        address _address, 
-        uint256 _percent, 
-        uint256 _claimed, 
-        uint256 _gClaimed, 
+        address _address,
+        uint256 _percent,
+        uint256 _claimed,
+        uint256 _gClaimed,
         uint256 _max
     ) public onlyOwner {
         require(terms[_address].max == 0, "address already exists");
-        terms[_address] = Term({
-            percent: _percent,
-            claimed: _claimed,
-            gClaimed: _gClaimed,
-            max: _max
-        });
+        terms[_address] = Term({percent: _percent, claimed: _claimed, gClaimed: _gClaimed, max: _max});
         require(totalAllocated.add(_percent) <= maximumAllocated, "Cannot allocate more");
         totalAllocated = totalAllocated.add(_percent);
     }
@@ -226,8 +222,8 @@ contract GenesisClaim is Ownable {
     /**
      * @notice all claims tracked under gClaimed (and track rebase)
      */
-     function treatAllAsStaked() external {
+    function treatAllAsStaked() external {
         require(msg.sender == dao, "Sender is not DAO");
         useStatic = false;
-     }
+    }
 }
